feat: implement user mutations
